+++ RELEASE_NOTES.md
@@ -1,3 +1,9 @@
+### 2.0.10
+- Expose DomHelpers.nodeStrShort, nodeStr, DomEdit
+- Added more `c` helpers (ac, hrc, tbodyc, olc)
+- Fix for Store.iter
+-
+
 ### 2.0.9
 - Remove cleanup diagnostics